refator(home): remove `feature` subdir & tweak wayland app configs
Also add `swapwindow` binds to hyprland. Previously, moving a window
would move it _into_ a stack. So with three windows open, two on the
left and one on the right, moving the top left window to the right would
make it so that there are now two windows on the right. This is
sometimes annoying when I want to give a window more space, because I
have to move the window into a position, and then move out the other
window that was already there, making me do extra work.

Now there's a separate binds for swapping windows, so a top left window
will swap with the right window and become the new sole right window,
sending the old right window to the top left.