feat(home/cli): tweak fd/rg ignores, install pistol previewer for fzf
I was having weird bugs with Telescope not producing result in Neovim
for some weird reason. Traced it down to my fd/rd ignore settings, even
though those applications worked just fine on the CLI directly.

At least Telescope works again now. I got kind of fed up with managing a
growing if/else expression for doing fzf previews, so I installed a
dedicated application for this. Its a go program that uses some
mime-type libraries. It's kind of like lesspipe. But already I'm having
to do exceptions to get the results I wants, so I'll have to see how far
I want to take this. I might be able to use it in Neomutt as well, so
there might be value to this.