feat: add plover - the open source stenography engine
This required some work. I needed to enable Appimage, Xwayland, add a
udev rule, add a group to my user, and add a flake.

Plover doesn't support Wayland out of the box, so I'm using an
unofficial flake that has integrated a pending PR for implementing
`uinput` (whatever that really is) in plover in order to support
Wayland.