feat(hyprland): set new window direction and no gaps for single windows
I missed these settings until now, having new windows always appear to
the right or bottom is such an improvement. By default, it positions
newly created windows based on the position of the cursor in the
existing window, which is super awkward when you have your cursor
disabled 99% of the time. `no_gaps_when_only` is also kind of nice,
previously I did it in a roundabout way with window rules, but this is
way cleaner.