fix(shell): hyprland binary name starts with capital letter
This is what was causing my "command not found" problem with newer
versions all this time...

Seriously, what maniac names their binaries with upper-case letters?
Seems like in version 0.43 and before, there was a lower-case symlink
for sane people, but it's been removed since? This is what caused my
launch-on-login command to break and all this time I thought
`hyprland` missing from path was a problem with the NixOS
module/package.