feat(systems/common): enable experimental pipe-operators 😁
These are, like, the nicest part of using a functional programming
language.

I don't get the resistance these got from the Nix maintainers for
"increasing accidental complexity". As if the average person getting
started with Nix never wrote Bash (recognize `|`, anyone?) before in
their lives. That when Nix is a strictly expression-based language,
which is arguably already a way bigger hurdle for newcomers to grasp.