refactor(system/raydesk): use hyprland & co from nixpkgs
I was using the hyprland and hyprsplit plugin flakes. I recently found
out this plugin is included in nixpkgs and was on the same version as I
had pinned anyway. So I don't feel like there is a reason to use the
flakes.

Up until now, it was rebuilding hyprland almost every time I tried to
update it, and I had version mismatches between hyprland and hyprsplit.