feat(system/raydesk): install qmk udev rules
Probably needed to detect a keeb controller in bootloader mode? Haven't
been able to get QMK flashing working on my NixOS desktop yet, I'm
missing `bootloadHID` and it's not in nixpkgs.

Nevertheless, if when I want to do something with QMK again in the
future, it's probably for the best to have these rules.