refactor: move home-manager config module definition out of flake.nix
I had this stupid `mkHomeManagerModule` helper for defining an inline
module for the home-manager configs used in each host. But now it's not
necessary any more because I realised that I could just put the
definition inside each host's `configuration.nix` and import the right
files based on the hostname.