feat(home/mail): update config to match `v1.1.0` himalaya module
The latest version of himalaya had some changes to it's config, and it
took a while for the updated home-manager module to be merged. Now the
himalaya module is compatible again with the `emails.accounts` settings.

Not quite using this yet, I'm waiting for `neverest` to become ready
since `mbsync` doesn't support Oauth.