docs(system/raymac): add comment reminding about driver kit for kmonad
I lost a day fiddling with this. Since kmonad has no nix-darwin support
I'm installing it as a flake and wrote a launch daemon for it. Only it
has a runtime dependency unique to MacOS which was still installed from
when I did it manually before I started using nix-darwin. So I was under
the impression that it just worked, but really it requires manual
intervention.