feat(system): enable `smartd` service for raydesk & raytop hosts
`smart` is a system for checking drive diagnostics. Smartd is a daemon
that performs these checks on a schedule as well as on first boot. I'd
like notifications of some sort to see the results, but I haven't
figured that out. Looking at journalctl tells me that one of my desktop
hard drives is failing.