refactor(home/common): use `home.sessionVariables` for `$NIX_LD` hack
This is a lot cleaner, instead of assigning and exporting the variable
inside the shell's config file. I hadn't spotted this until now, my
understanding of Nix configuration has improved since I first included
this tidbit.

I tried to think why the original tip was written like this, as a nix
expression executed on launching interactive shells, but I can't figure
it out. I checked, and the variable merely contains a store path. So
this nix expression is reading a store path from a file that's shipped
with the `pkgs.stdenv.cc` package. I don't see why this has to happen on
shell initialization, rather than upon rebuilding the system config. I
figure that maybe when using channels, the nixpkgs version can be
updated without rebuilding the config? So after updating the channel,
the `$NIX_LD` variable would be pointing at the wrong path? In my case,
the path cannot change without rebuilding the system config, which would
also update variable when set through `home.sessionVariables`, so I
think this is way cleaner and behaves the same.