feat(hosts/raymac): rework kmonad config & swap to kanata
- Friendship with kmonad over. Kanata is my new friend.
- Chords/QMK Combos!
- Tap-holds now allow repeating tap keys, by tapping-releasing-holding
  (just like on keyd/QMK)
- tap-hold on the pinkies and ring fingers now auto-resolve to taps for
  specific nearby keys that are frequently rolled together.

kanata is a kmonad clone in Rust. One significant improvement is that it
has chords/QMK combos, which I something I was sorely missing in kmonad.
Also has slightly more advanced top-hold configuration. I moved to
kanata, because an update had once again left me with a non-responding
keyboard, which took a whole weekend to troubleshoot. Turns out that
kmonad moved to a higher version of dext (driverkit-extension), which
required me to manually update the dext, and add a new launchDaemon for
the driverkit. Then the driverkit binary _also_ needed "Input Monitor"
privileges.

This commit includes the changes to my kmonad config, for posterity.