feat(home/common): rework nushell fzf history command
Query the sqlite db directly, instead of piping the `history` command,
as this is way faster, as per casual testing with the `timeit {}`
command. Also de-duplicates the results.

There is more data available to query, I'm curious as to in what ways
the history widget can be made even smarter.