feat(home/common/librewolf): open external links in a new window
It used to be the case that opening a link in Discord for example, would
open a tab in the most recently active browser window. But often that
window is on an inactive desktop workspace, so I would never see it
appear and wonder what happened with the link I just clicked on.