perf(home/common/nu): no config & stdlib `fzf` integration commands
Nu makes it trivial to benchmark/profile using `timeit {}` and `debug
profile`. From testing, it is evident that not sourcing the config and
the stdlib makes startup time of a subshell a lot faster, and the `fzf`
commands still work without, because the variables are expanded by the
parent shell.