feat(home/common/script): add nu script for fetching new mail
Takes a list of account names. Uses all configured accounts by default.
Accounts are synced in parallel (but no multiple mbsync instances for
the same account are allowed). Produces a list of new mail, and sends a
desktop notification. Uses `notify-send --action` option to choose a
folder with new mail to open in a new aerc window immediately.