fix(home/common/mail): shrunken unicode icons in aerc
It's a double-width unicode thing, I believe. Not sure, unicode symbols
are weird and complicated. I remember that `st` (the suckless terminal)
had different patches for how to deal with symbols. In the case of
kitty, which I'm using nowadays, it seems it shrinks icons to fit inside
a single character cell, unless it's followed by a space? In which case
it's allowed to overflow to the neighbouring cell. So adding a space
after all the icons makes them appear properly sized and proportioned,
which is a lot nicer. The envelope listings still look a little weird
when there is a bunch of icons on the same mail, but at least they're
consistent.