feat(home/common/dev): install `make`
I had noticed for since like forever that `telescope-fzf-native.nvim`
was disabled. For some reason, I never felt like investigating that. For
all this time, I had a check for whether `make` was installed as a part
of this plugin's `enable` option.

I haven't For the majority of time using NixOS had `make` installed,
because I haven't been running Suckless software anymore.

So all this time it was disabled, and I could have easily fixed it...