fix(home/raytop): disable xwayland scaling in hyprland to fix blurriness
I don't get what it thinks it's scaling. On my desktop it doesn't do
this. I thought it was Electron being bad and causing Vesktop to be
unusably blurry. Turns out it was Xwayland under Hyprland that was
causing it. As far as I'm aware, I haven't configured any scaling
though.