feat(home/common): install pandoc and typst
I've been using pandoc for a long time, it makes sense to include it in
my system.

Typst is the new LaTeX. If I ever write a full-blown report again, I'm
unlikely to reach for LaTeX any more. Unfortunately, Typst isn't fully
up to snuff as a PDF engine for pandoc.