feat(home/common): better screenshot setup for hyprland
Uses grimblast - a more robust wrapper around grim & slurp, and satty -
a very nice looking and simple image markup tool for drawing and
cropping images.
Temporarily disables window transparency and freezes the screen when
taking screenshots, allows for quick highlights/redactions, and copying
the result to the clipboard OR saving to the tmp dir and drag-n-dropping
with ripdrag.