fix(home/common): move `pwmenu` command from common hyprland into raytop
I already use meta+ctrl+m as a keybind for cmus, so this turned out to
be a conflicting keybind. I don't keep music on my laptop, so it's fine
to use there. I also don't control my volume on my desktop in software,
as my stereo set has a volume knob, so I don't really need `pwmenu`
their anyway. This returns the config to the old state where the volume
menu command is specific to my laptop.